After splitting of a branch (r4199) including code for moving from wicket 1.4 to wicket 6, some more modifications in the direction of that target.

+++ vlo_web_app/src/main/java/eu/clarin/cmdi/vlo/pages/FacetedSearchPage.java
@@ -270,7 +270,7 @@ public class FacetedSearchPage extends BasePage {
 
                 int offset = 0;
                 int fetchSize = 1000;
-                int totalResults = dataProvider.size();
+                int totalResults = (int) dataProvider.size();
                 while (offset < totalResults) {
                     Iterator<SolrDocument> iter = dataProvider.iterator(offset, fetchSize);
                     while (iter.hasNext()) {
